Move process to the new CLI
